Conflicts:
	android-application/ChineseCheckers/libs/cc-api.jar
	cc-api-lib/target/cc-api.jar
	cc-api-lib/target/maven-archiver/pom.properties
	server/server-application/backend-api-server/conf/routes
	server/server-application/backend-api-server/lib/cc-api.jar

+++ android-application/ChineseCheckers/src/main/java/ca/brocku/chinesecheckers/network/spice/requests/JoinGameRequest.java
@@ -14,9 +14,9 @@ import ca.brocku.chinesecheckers.network.spice.ApiRequest;
  */
 public class JoinGameRequest extends ApiRequest<JoinGameReceivable> {
     private int playerCount;
-    private int userId;
+    private long userId;
 
-    public JoinGameRequest(int userId, int playerCount) {
+    public JoinGameRequest(long userId, int playerCount) {
         super(JoinGameRequest.class);
 
         this.userId = userId;
